+++ src/Program.cs
@@ -9,11 +9,13 @@ using AzureMcp.Models;
 using AzureMcp.Models.Command;
 using AzureMcp.Services.Azure.AppConfig;
 using AzureMcp.Services.Azure.Cosmos;
+using AzureMcp.Services.Azure.KeyVault;
 using AzureMcp.Services.Azure.Kusto;
 using AzureMcp.Services.Azure.Monitor;
 using AzureMcp.Services.Azure.Postgres;
 using AzureMcp.Services.Azure.ResourceGroup;
 using AzureMcp.Services.Azure.Search;
+using AzureMcp.Services.Azure.ServiceBus;
 using AzureMcp.Services.Azure.Storage;
 using AzureMcp.Services.Azure.Subscription;
 using AzureMcp.Services.Azure.Tenant;
@@ -74,6 +76,8 @@ internal class Program
         services.AddSingleton<IAppConfigService, AppConfigService>();
         services.AddSingleton<ISearchService, SearchService>();
         services.AddSingleton<IPostgresService, PostgresService>();
+        services.AddSingleton<IKeyVaultService, KeyVaultService>();
+        services.AddSingleton<IServiceBusService, ServiceBusService>();
         services.AddSingleton<CommandFactory>();
     }
 }