+++ CHANGELOG.md
@@ -1,16 +1,24 @@
 # Release History
 
-## 0.0.14 (Unreleased)
+## 0.0.15 (Unreleased)
 
 ### Features Added
 
-- Support for Kusto (Azure Data Explorer).  https://github.com/Azure/azure-mcp/pull/21
+- Support for getting properties and runtime information for Azure Service Bus queues, topics, and subscriptions. https://github.com/Azure/azure-mcp/pull/150/
+- Support for peeking at Azure Service Bus messages from queues or subscriptions. https://github.com/Azure/azure-mcp/pull/144
 
 ### Breaking Changes
 
 ### Bugs Fixed
 
 ### Other Changes
+
+## 0.0.14 (2025-05-07)
+
+### Features Added
+
+- Support for Azure Key Vault keys https://github.com/Azure/azure-mcp/pull/119
+- Support for Kusto (Azure Data Explorer).  https://github.com/Azure/azure-mcp/pull/21
 
 ## 0.0.13 (2025-05-06)